pyproject.toml
[project]
name = "santander-notification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "pandas",
    "requests>=2.32.3",
    "google-auth>=2.38.0",
    "google-api-python-client>=2.165.0",
    "google-auth-oauthlib>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "types-requests>=2.31.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["santander_notification"]
